Use environment variable to specify Dockerfile
I added some metadata to previous versions so we can easily track which
dockerfile was used for the build

I renamed dockerfiles so they all start with 'unity', makes it easier to
follow.

I added latest versions and I'm trying a fix for #19 even tho the issue
was closed, the problem happens for latest 2017 versions which may be a
problem for adoption of the CI and everyone deserves a CI no matter its
unity version.

I also removed some WET code in the .gitlab-ci.yml jinja template. So
now, everything use the same docker build commands (passing variables
even if they're not always all used).