WIP: Add check_new_version.py with tests
This is kind of cool, not the best script I wrote so far, but it gets
the job done. These scripts deserve some refactor, but the cool thing is
that the combination of both check_new_version.py and gitlab_ci_generator.py
makes me run two commands and voilà, it starts building missing versions
grabbed from that magic unity url.